requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-WTF==1.1.1
Werkzeug==2.3.7
email-validator==2.0.0
python-dotenv==1.0.0
# Note: psycopg2 requires Microsoft Visual C++ Build Tools
# Download from: https://visualstudio.microsoft.com/visual-cpp-build-tools/
# psycopg2==2.9.9
azure-cosmos==4.5.1
azure-storage-blob==12.19.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
gunicorn==21.2.0
Flask-Migrate==4.0.5
Flask-Limiter==3.5.0
Flask-Talisman==1.1.0 